+++ cli/src/main/kotlin/tech/coner/trailer/cli/command/webapp/WebappCommand.kt
@@ -1,16 +1,8 @@
 package tech.coner.trailer.cli.command.webapp
 
-import com.github.ajalt.clikt.parameters.groups.cooccurring
-import com.github.ajalt.clikt.parameters.options.option
-import com.github.ajalt.clikt.parameters.options.required
-import com.github.ajalt.clikt.parameters.types.enum
 import org.kodein.di.DI
-import org.kodein.di.instance
 import tech.coner.trailer.cli.command.BaseCommand
 import tech.coner.trailer.cli.command.GlobalModel
-import tech.coner.trailer.io.Webapp
-import tech.coner.trailer.io.service.ConfigurationService
-import tech.coner.trailer.webapp.results.resultsWebapp
 
 class WebappCommand(
     di: DI,
