+++ io-test/src/integration-test/kotlin/tech/coner/trailer/io/util/ServiceContainer.kt
@@ -1,6 +1,14 @@
 package tech.coner.trailer.io.util
 
-import org.kodein.di.*
+import java.nio.file.Path
+import kotlin.io.path.createDirectories
+import kotlinx.coroutines.runBlocking
+import org.kodein.di.DI
+import org.kodein.di.DIAware
+import org.kodein.di.diContext
+import org.kodein.di.direct
+import org.kodein.di.instance
+import org.kodein.di.on
 import tech.coner.trailer.datasource.snoozle.ConerTrailerDatabase
 import tech.coner.trailer.di.EnvironmentHolder
 import tech.coner.trailer.io.Configuration
@@ -9,10 +17,22 @@ import tech.coner.trailer.io.TestEnvironments
 import tech.coner.trailer.io.WebappConfiguration
 import tech.coner.trailer.io.di.ioTestKodeinDi
 import tech.coner.trailer.io.payload.ConfigAddDatabaseParam
-import tech.coner.trailer.io.service.*
-import java.nio.file.Path
-import kotlin.io.path.createDirectories
-import kotlinx.coroutines.runBlocking
+import tech.coner.trailer.io.service.ClubService
+import tech.coner.trailer.io.service.ConfigurationService
+import tech.coner.trailer.io.service.CrispyFishClassService
+import tech.coner.trailer.io.service.CrispyFishEventMappingContextService
+import tech.coner.trailer.io.service.EventService
+import tech.coner.trailer.io.service.MotorsportRegEventService
+import tech.coner.trailer.io.service.MotorsportRegImportService
+import tech.coner.trailer.io.service.MotorsportRegMemberService
+import tech.coner.trailer.io.service.MotorsportRegPeopleMapService
+import tech.coner.trailer.io.service.ParticipantService
+import tech.coner.trailer.io.service.PersonService
+import tech.coner.trailer.io.service.PolicyService
+import tech.coner.trailer.io.service.RankingSortService
+import tech.coner.trailer.io.service.RunService
+import tech.coner.trailer.io.service.SeasonPointsCalculatorConfigurationService
+import tech.coner.trailer.io.service.SeasonService
 
 class ServiceContainer(
     val root: Path,
@@ -33,10 +53,9 @@ class ServiceContainer(
         ),
         defaultDatabaseName = databaseConfiguration.name,
         webapps = Configuration.Webapps(
-            results = WebappConfiguration(
+            competition = WebappConfiguration(
                 port = 8080,
-                exploratory = true,
-                wait = true
+                exploratory = true
             )
         )
     )
