+++ io/src/main/kotlin/tech/coner/trailer/io/WebappConfiguration.kt
@@ -4,6 +4,5 @@ import com.fasterxml.jackson.annotation.JsonIgnore
 
 data class WebappConfiguration(
     val port: Int,
-    @JsonIgnore val exploratory: Boolean,
-    @JsonIgnore val wait: Boolean = true
+    @JsonIgnore val exploratory: Boolean
 )
