+++ webapp-competition/src/main/kotlin/exploratory/resource/HelloResource.kt
