+++ webapp-competition/src/main/node/.gitignore
